+++ cmd/ctr/main.go
@@ -6,6 +6,7 @@ import (
 	"log"
 	"os"
 
+	"github.com/containerd/containerd/cmd/ctr/commands/plugins"
 	versionCmd "github.com/containerd/containerd/cmd/ctr/commands/version"
 	"github.com/containerd/containerd/namespaces"
 	"github.com/containerd/containerd/server"
@@ -65,6 +66,8 @@ containerd CLI
 		},
 	}
 	app.Commands = append([]cli.Command{
+		plugins.Command,
+		versionCmd.Command,
 		applyCommand,
 		containersCommand,
 		contentCommand,
@@ -81,8 +84,6 @@ containerd CLI
 		runCommand,
 		snapshotCommand,
 		tasksCommand,
-		pluginsCommand,
-		versionCmd.Command,
 	}, extraCmds...)
 	app.Before = func(context *cli.Context) error {
 		if context.GlobalBool("debug") {
